refactor(26237): fix mocks
