+++ hivemq-edge/src/frontend/src/__test-utils__/rjsf/schema.mocks.ts
@@ -1,70 +1,5 @@
 import { RJSFSchema } from '@rjsf/utils'
 
-export const GENERATE_DATA_MODELS = (short = false, title?: string): RJSFSchema => {
-  const model: RJSFSchema = {
-    title: title || 'A registration form',
-    description: 'A simple form example.',
-    type: 'object',
-    required: ['firstName', 'lastName'],
-    properties: {
-      firstName: {
-        type: 'string',
-        title: short ? 'First String' : 'firstName',
-        default: 'Chuck',
-        examples: 'firstName',
-      },
-      lastName: {
-        type: 'string',
-        title: short ? 'Second String' : 'lastname',
-      },
-      age: {
-        type: 'integer',
-        title: short ? 'Number' : 'age',
-      },
-      subItems: {
-        type: 'object',
-        title: 'subItems',
-        examples: 'subItems',
-
-        properties: {
-          name: {
-            examples: 'name',
-            type: 'string',
-            title: 'name',
-            default: 'Default name',
-          },
-          type: {
-            type: 'string',
-            title: 'type',
-            default: 'Default type',
-          },
-        },
-      },
-    },
-  }
-
-  if (!short)
-    model.properties = {
-      ...model.properties,
-      bio: {
-        type: 'string',
-        title: 'bio',
-      },
-      password: {
-        type: 'string',
-        title: 'password',
-        minLength: 3,
-      },
-      telephone: {
-        type: 'string',
-        title: 'telephone',
-        minLength: 10,
-      },
-    }
-
-  return model
-}
-
 export const MOCK_MQTT_SCHEMA_REFS: RJSFSchema = {
   definitions: {
     address: {
